Modifying script for simpler version of test case modifications
