+++ Cargo.toml
@@ -6,14 +6,13 @@ rust-version = "1.76"
 default-run = "visualizer"
 
 [dependencies]
-clap = { version = "4.5.31", features = ["derive"], optional = true }
+clap = { version = "4.5.31", features = ["derive"]}
 itertools = "0.14.0"
 ordered-float = "5.0.0"
 random_color = { version = "1.0.0", optional = true }
 rerun = { version = "0.22.1", optional = true }
 serde = { version = "1.0.218", features = ["derive"] }
 serde_json = "1.0.140"
-struct-field-names-as-array = "0.3.0"
 walkdir = "2.5.0"
 
 [dev-dependencies]
@@ -25,7 +24,7 @@ rstest_reuse = "*"
 tempfile = "3.17.1"
 
 [features]
-visualizer = ["clap", "random_color", "rerun"]
+visualizer = ["random_color", "rerun"]
 
 [[bin]]
 name = "generate-rotated-ipa-polygons"
@@ -37,8 +36,8 @@ path = "src/bin/run_visualizer.rs"
 required-features = ["visualizer"]
 
 [[bin]]
-name = "sync-test-case-attrs"
-path = "src/bin/sync_test_case_attributes.rs"
+name = "modify-test-case-fields"
+path = "src/bin/modify_test_case_fields.rs"
 
 [[bench]]
 name = "convex_hull"
